refact: commentlist commentview를 분리
