--- Mapping/Comment/CommentViewModel.swift
@@ -1,18 +1,0 @@
-//
-//  CommentViewModel.swift
-//  Mapping
-//
-//  Created by 김민정 on 12/31/24.
-//
-
-import SwiftUI
-
-struct CommentViewModel: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    CommentViewModel()
-}
